Refactors API test cases
