+++ noq_django/noq_django/api/test/test_products_api.py
@@ -3,25 +3,16 @@ sys.path.append("....backend") # Adds folder where backend is to python modules 
 import json
 from django.test import TestCase
 from django.contrib.auth.models import User, Group
-from backend.models import Host, Product, Region
-from django.test import Client as TestClient
+from backend.models import Host, Client, Product, Region
+from django.test import Client
 from ..host_api import router
+from .test_data import TestData
 
 class TestProductsApi(TestCase):
-    user = None
-    client = None
-    username = "host"
-    password = "hostpassword"
-
-    def host_login(self):
-        # Create host user for the tests if not existing
-        if User.objects.filter(username=self.username).first() == None:
-            self.user = User.objects.create_user(username=self.username, password=self.password)
-            group_obj, created = Group.objects.get_or_create(name="host")
-            self.user.groups.add(group_obj)
+    t_data = None
 
         # Login the host user
-        self.client = TestClient(router)
+        self.client = Client(router)
         self.client.login(username=self.username, password=self.password)
 
     def delete_user(self):
@@ -36,33 +27,27 @@ class TestProductsApi(TestCase):
         Region.objects.all().delete()
 
     def setUp(self):
+        self.t_data = TestData()
         # log in host user for the tests
-        self.host_login()
-        # Create products that can be used during the tests
-        if Product.objects.filter(id=1).first() == None:
-            region = Region.objects.create(name="Malmö")
-            hostA = Host.objects.create(name="Host 1", street="", postcode="", city="Malmö", region_id=region.id)
-            hostB = Host.objects.create(name="Host 2", street="", postcode="", city="Malmö", region_id=region.id)
-            productA = Product.objects.create(name="room", total_places=2, host_id=hostA.id, type="room")
-            productB = Product.objects.create(name="room", total_places=5, host_id=hostB.id, type="room")
-            productC = Product.objects.create(name="woman-only", total_places=3, host_id=hostB.id, type="woman-only")
+        self.t_data.user_login(user_group="host", nr_of_users=1)
 
 
     def test_list_all_products(self):
         products_count = Product.objects.all().count()
-        response = self.client.get("/api/host/products")
+        response = self.t_data.test_client.get("/api/host/products")
 
         self.assertEqual(response.status_code, 200)
 
         data = json.loads(response.content)
         count = len(data)
         self.assertEqual(count, products_count)
+
 
     def test_list_host_products(self):
         host = Host.objects.all().first()
         products_count = Product.objects.filter(host_id=host.id).count()
         url = "/api/host/hosts/" + str(host.id) + "/products"
-        response = self.client.get(url)
+        response = self.t_data.test_client.get(url)
 
         self.assertEqual(response.status_code, 200)
 
@@ -70,10 +55,11 @@ class TestProductsApi(TestCase):
         count = len(data)
         self.assertEqual(count, products_count)
 
+
     def test_get_unique_product(self):
         product = Product.objects.all().first()
         url = "/api/host/products/" + str(product.id)
-        response = self.client.get(url)
+        response = self.t_data.test_client.get(url)
 
         self.assertEqual(response.status_code, 200)
 
@@ -108,12 +94,13 @@ class TestProductsApi(TestCase):
             },
             "type": "room"
         }
-        response = self.client.post(url, data, content_type='application/json')
+        response = self.t_data.test_client.post(url, data, content_type='application/json')
 
         self.assertEqual(response.status_code, 201)
 
         self.assertEqual(Product.objects.all().count(), expected_count)
     '''
+
 
     def test_update_product(self):
         product = Product.objects.all().first()
@@ -141,12 +128,13 @@ class TestProductsApi(TestCase):
             "type": "room"
         }
         headers = {'Content-Type': 'application/json', 'accept': 'application/json'}
-        response = self.client.put(url, json.dumps(data), headers=headers)
+        response = self.t_data.test_client.put(url, json.dumps(data), headers=headers)
 
         self.assertEqual(response.status_code, 200)
 
         product = Product.objects.all().first()
         self.assertEqual(product.total_places, new_total_places)
+
 
     def test_delete_product(self):
         product = Product.objects.all().first()
@@ -154,14 +142,15 @@ class TestProductsApi(TestCase):
 
         url = "/api/host/products/" + str(product.id)
         # data should be product_id, not sure how to use parameter
-        response = self.client.delete(url, product_id=id)
+        response = self.t_data.test_client.delete(url, product_id=id)
 
         self.assertEqual(response.status_code, 200)
 
         product = Product.objects.all().first()
         self.assertNotEqual(product.id, id)
 
+
     def tearDown(self):
         # After the tests delete all data generated for the tests
-        self.delete_user()
-        self.delete_products()
+        self.t_data.delete_users()
+        self.t_data.delete_products()
