+++ noq_django/noq_django/api/test/test_user_api.py
@@ -2,181 +2,27 @@ import sys
 sys.path.append("....backend")
 import json
 from django.test import TestCase
-from django.db.models import Q
 from django.contrib.auth.models import User, Group
 from backend.models import (Host, Client, Product, Region, Booking,
                             Available, State, BookingStatus)
-from django.test import Client as TestClient
 from datetime import datetime, timedelta
-from ..host_api import router
+from .test_data import TestData
 
 class TestProductsApi(TestCase):
-    user = None
-    client = None
-    username = "user.user@test.nu"
-    username_two = "user.user.two@test.nu"
-    password = "userpassword"
-
-    def user_login(self):
-        # Create users for the tests if not existing
-        if User.objects.filter(username=self.username).first() == None:
-            self.user = User.objects.create_user(
-                username=self.username, password=self.password)
-            group_obj, created = Group.objects.get_or_create(name="user")
-            self.user.groups.add(group_obj)
-
-            self.user = User.objects.create_user(
-                username=self.username_two, password=self.password)
-            group_obj, created = Group.objects.get_or_create(name="user")
-            self.user.groups.add(group_obj)
-
-        # Login the host user
-        self.test_client = TestClient(router)
-        self.test_client.login(
-            username=self.username, password=self.password)
-
-
-    def create_clients(self, users, region):
-        for user in users:
-            client = Client.objects.create(
-                first_name=user.username, gender="K", user=user, region=region)
-
-
-    def bookings_count_per_date(self, product):
-        # Return count for bookings for one week including current date
-        start_date = datetime.now().date()
-        time_period_in_days = 7
-        booking_counts = {}
-
-        for i in range(time_period_in_days):
-            date_time = start_date + timedelta(days=i)
-            count = Booking.objects.filter(
-                Q(product=product)
-                & Q(start_date__lte=date_time)
-                & Q(end_date__gt=date_time)
-                & ~Q(status=State.DECLINED)
-                & ~Q(status=State.IN_QUEUE)
-            ).count()
-            date = f"{date_time:%Y-%m-%d}"
-            booking_counts[date] = count
-
-        return booking_counts
-
-
-    def calc_available(self, product):
-        bookings_per_day = self.bookings_count_per_date(product)
-        for date in bookings_per_day:
-            places_left = product.total_places - bookings_per_day[date]
-
-            existing_availability = Available.objects.filter(
-                product=product, available_date=date
-            ).first()
-
-            if existing_availability:
-                existing_availability.places_left = places_left
-                existing_availability.save()
-            else:
-                product_available = Available(
-                    available_date=date,
-                    product=Product.objects.get(id=product.id),
-                    places_left=places_left,
-                )
-                product_available.save()
-
-
-    def delete_user(self):
-        # Delete the user created for the tests
-        user = User.objects.get(username=self.username)
-        user.delete()
-
-
-    def delete_products(self):
-        # Delete all products created for the tests
-        Product.objects.all().delete()
-        Available.objects.all().delete()
-        Host.objects.all().delete()
-        Region.objects.all().delete()
-
-
-    def add_booking_statuses(self):
-        statuses = [
-            {"id": State.PENDING, "description": "pending"},
-            {"id": State.DECLINED, "description": "declined"},
-            {"id": State.ACCEPTED, "description": "accepted"},
-            {"id": State.CHECKED_IN, "description": "checked_in"},
-            {"id": State.IN_QUEUE, "description": "in_queue"},
-            {"id": State.RESERVED, "description": "reserved"},
-            {"id": State.CONFIRMED, "description": "confirmed"},
-        ]
-
-        for status in statuses:
-            if not BookingStatus.objects.filter(id=status["id"]).exists():
-                booking_status = BookingStatus.objects.create(
-                    id=status["id"], description=status["description"]
-                )
-                booking_status.save()
-
+    t_data = None
 
     def setUp(self):
+        # Add data to the db
+        self.t_data = TestData()
         # log in host user for the tests
-        self.user_login()
-        # Create products that can be used during the tests
-        if Product.objects.filter(id=1).first() == None:
-            region = Region.objects.create(name="Malmö")
-
-            hostA = Host.objects.create(
-                name="Host A",
-                street="Bennets Väg 9",
-                postcode="21367",
-                city="Malmö",
-                region_id=region.id
-            )
-            hostB = Host.objects.create(
-                name="Host B",
-                street="Idunavägen 11",
-                postcode="21619",
-                city="Malmö",
-                region_id=region.id
-            )
-
-            for i in range(4):
-                nr_of_places = 2 * (i + 1)
-                host_id = hostA.id if i < 2 else hostB.id
-                product = Product.objects.create(
-                    name="room",
-                    total_places=nr_of_places,
-                    host_id=host_id,
-                    type="room"
-                )
-
-            product = Product.objects.create(
-                name="woman-only",
-                total_places=1,
-                host_id=hostA.id,
-                type="woman-only"
-            )
-
-        # Create client
-        users = [
-            User.objects.get(username="user.user@test.nu"),
-            User.objects.get(username="user.user.two@test.nu")
-        ]
-        self.create_clients(users, region)
-
-
-        # Count availability for a week
-        for product in Product.objects.all():
-            self.calc_available(product)
-
-        # Add booking statuses
-        self.add_booking_statuses()
+        self.t_data.user_login(user_group="user", nr_of_users=2)
 
 
     def test_get_available_products(self):
         # No Bookings, so expects to get five products (3 Host A, 2 Host B)
         current_date = datetime.now().date()
         url = "/api/user/available/" + str(current_date)
-        response = self.test_client.get(url)
+        response = self.t_data.test_client.get(url)
 
         self.assertEqual(response.status_code, 200)
         data = json.loads(response.content)
@@ -211,7 +57,7 @@ class TestProductsApi(TestCase):
         }
 
         url = "/api/user/request_booking"
-        response = self.test_client.post(url, data, content_type='application/json')
+        response = self.t_data.test_client.post(url, data, content_type='application/json')
 
         # Check status code is 200 and the status of the booking is pending
         self.assertEqual(response.status_code, 200)
@@ -240,7 +86,7 @@ class TestProductsApi(TestCase):
         product = Product.objects.get(total_places=1)
 
         client = Client.objects.get(
-            user=User.objects.get(username="user.user.two@test.nu"))
+            user=User.objects.get(username=self.t_data.usernames[1]))
 
         booking = Booking(
             start_date=start_date,
@@ -262,7 +108,7 @@ class TestProductsApi(TestCase):
         }
 
         url = "/api/user/request_booking"
-        response = self.test_client.post(url, data, content_type='application/json')
+        response = self.t_data.test_client.post(url, data, content_type='application/json')
 
         # Check status code is 200 and the bookings count matches
         self.assertEqual(response.status_code, 200)
@@ -285,5 +131,5 @@ class TestProductsApi(TestCase):
 
     def tearDown(self):
         # After the tests delete all data generated for the tests
-        self.delete_user()
-        self.delete_products()
+        self.t_data.delete_users()
+        self.t_data.delete_products()
